change config format
