+++ src/main/java/io/th0rgal/oraxen/mechanics/provided/gameplay/furniture/FurnitureListener.java
@@ -12,8 +12,6 @@ import io.th0rgal.oraxen.utils.BlockHelpers;
 import io.th0rgal.oraxen.utils.Utils;
 import io.th0rgal.oraxen.utils.breaker.BreakerSystem;
 import io.th0rgal.oraxen.utils.breaker.HardnessModifier;
-import io.th0rgal.oraxen.utils.limitedplacing.AllowPlacingOn;
-import io.th0rgal.oraxen.utils.limitedplacing.DenyPlacingOn;
 import io.th0rgal.oraxen.utils.limitedplacing.LimitedPlacing;
 import io.th0rgal.protectionlib.ProtectionLib;
 import org.bukkit.*;
@@ -160,14 +158,14 @@ public class FurnitureListener implements Listener {
         if (mechanic == null || !mechanic.hasLimitedPlacing()) return;
 
         LimitedPlacing limitedPlacing = mechanic.getLimitedPlacing();
-        AllowPlacingOn allowPlacingOn = limitedPlacing.getAllowedPlacing();
-        DenyPlacingOn denyPlacingOn = limitedPlacing.getDeniedPlacing();
         Block placedAgainst = block.getRelative(event.getBlockFace()).getRelative(BlockFace.DOWN);
 
-        if (allowPlacingOn != null && !allowPlacingOn.checkAllowedMechanic(placedAgainst)) {
-            event.setCancelled(true);
-        } else if (denyPlacingOn != null && denyPlacingOn.checkDeniedMechanic(placedAgainst)) {
-            event.setCancelled(true);
+        if (limitedPlacing.getType() == LimitedPlacing.LimitedPlacingType.ALLOW) {
+            if (!limitedPlacing.checkLimitedMechanic(placedAgainst))
+                event.setCancelled(true);
+        } else if (limitedPlacing.getType() == LimitedPlacing.LimitedPlacingType.DENY) {
+            if (limitedPlacing.checkLimitedMechanic(placedAgainst))
+                event.setCancelled(true);
         }
     }
